LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-multicol-003-expected.txt










































PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="95"><span></span></div>
offsetLeft expected 95 but got 250
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-x="30"><span></span></div>
offsetLeft expected 30 but got 25
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="135"><span></span></div>
offsetLeft expected 135 but got 50
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
